:recycle: refactor - Ajuste no CRUD / UsuarioService na Controller /

+++ src/main/java/com/br/educonnect/persistence/repository/UserRepository.java
@@ -5,6 +5,6 @@ package com.br.educonnect.persistence.repository;
 import com.br.educonnect.persistence.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Usuario,Integer> {
+public interface UserRepository extends JpaRepository<Usuario,Long> {
 
 }
